Renamed the multiply function module
