+++ mul_table_func.py
